+++ tests/test_add.c
@@ -99,7 +99,7 @@ test_error_1(void)
   int err;
   cmat_t* m;
 
-  double v[] = {
+  float v[] = {
     1, 2,
     3, 4
   };
@@ -119,7 +119,7 @@ test_error_2(void)
   int err;
   cmat_t* m;
 
-  double v[] = {
+  float v[] = {
     1, 2,
     3, 4
   };
