+++ tests/test_destroy.c
@@ -23,7 +23,7 @@ test_normal_2(void)
 {
   int err;
   cmat_t* m;
-  double v[] = {
+  float v[] = {
     1, 2, 3,
     4, 5, 6,
   };
